+++ CHANGELOG.rst
@@ -6,7 +6,8 @@ Version 0.2.1
 =============
 
 * Parse.metadata() to return a DataFrame with all parsed pieces' metadata
-* improved the information that Parse objects return about themselves
+* Parse.get_labels() to retrieve labels of a particular kind
+* Parse.info() has improved the information that objects return about themselves
 * improved expansion of DCML harmony labels
 
 Version 0.2.0